pyproject.toml
[tool.poetry]
name = "warp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
jishaku = "^1.19.1"
kitsu-wrapper = "^1.0.0"
Pillow = "^7.2.0"
"nekos.py" = "^1.0.3"
psutil = "^5.7.2"
PyNaCl = "^1.4.0"
google-api-python-client = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
